Cmd: Add distinct reporting of no checksum response
In the current send_buffer() implementation, used for sending out powerline
signals from a command line, if a wait for checksum timeout occures, it is not
reported, even in verbose mode, other than incorrectly as a wrong checksum.
This could effectively hide this potential source of other heyu problems,
like "Unable to lock ..." events, from users and developers attention for
years.

This patch adds a distinct warning message in order to get users and
developers attention to these rare, but harmful events. To not force users
runnig commands in verbose mode, the events are reported unconditionally to
the syslog as well.

Created and tested against heyu-2.9.0.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>