Cmd: Limit the wait for checksum timeout to 1 second
In the current send_buffer() implementation, used for sending out powerline
signals from a command line, the same timeout value of 10s, as forced by the
calling functions, is used when waiting for both a checksum and READY
response.

It has been observed that a checksum response is either received immediatelly
or not at all. Then, waiting for 10 seconds seems useless, and even harmfull
since a heyu write lock is kept, preventing other commands from being run.

The patch limits the wait for checksum timeout to 1 second, that seems way
enough.

Created and tested against heyu-2.9.0.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>