Engine: provide check4poll() with checksums
In order to teach the engine how to recognize checksum bytes related to
outgoing CM11 commands, including powerline signals, found in the byte stream
read from the spool file, the check4poll() function must get access to
checksum values calculated for the relevant command sequences.

This patch allows for this by modifying the API of the identify_sent()
function, called by the check4poll(), adding a checksum, that is already
computed inside the identify_sent() body, to the list of its returned
variables.

Created and briefly tested against heyu-2.9.0.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>