Avoid launching a script if wrong script name specified
If a user happens to put a "LAUNCHER name ..." clause in his x10config
without providing a corresponding "SCRIPT -l name ..." definition,
a first script, in the x10config file order, is launched on every signal
that matches launch conditions of the inconsistent LAUNCHER.

This patch fixes the problem by preventing heyu from starting up
if such error in the x10config file is detected.

Created and tested against heyu-2.9.1.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>