extend rcs thermostat functions
With this patch applied, Heyu should try to decode and display three
more message types that can preceed temperature data reports: "Outside
Temperature", "Heat Setpoint" and "Cooling Setpoint". These labels
should be displayed in both the Heyu monitor output as well as the Heyu
engine log, but not from the rcs_req command line (only the temperature
is displayed if received on time, like it was before).

Reported-by: Brandt Daniels <PompRocker@aol.com>
Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>