Use microsleep(ENGINE_POLL) instead of sleep(1)
Now that we stopped using select() as a *sleep() replacement on linux,
both the Heyu engine and monitor fall back to polling the spool file
once a second while idle, irrespective of the ENGINE_POLL value, either
default or redefined in an x10config file.

In order to keep the ENGINE_POLL directive effective on linux, replace
the fallback arbitrary sleep(1) with an already defined custom
microsleep() function call, which is supposed to be portable.

This should make the ENGINE_POLL being respected on other platforms
as well, which have never been build with HASSELECT defined.

Created and tested against heyu-2.9.2.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>