Fix 'changed' handling for rain sensors
It had been reported that the ORE_CHGBITS_RAINRATE directive didn't
work as expected, not preventing consecutive oreRainRate signals
with close or even the same values from showing up in the Heyu log.
Further investigation made it clear that the problem was rather the
'changed' state calculated incorrectly for rain sensors, no matter how
the ORE_CHGBITS_RAINRATE was set.

Examination of the Heyu code revealed that the source of this problem
was incorrect offset used for accessing oreRainTot data in the data
store. As a result, oreRainRate data stored for future reference was
partially overwritten by oreRainTot, making correct comparison with
previous values required for 'changed' evaluation not possible.

The patch is supposed to correct the problem.

Created against Heyu 2.9.3.
Fixes #3

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>