Fix 0xff checksum before triple 0xff mark case
It may happen that a checksum of the 0xff value is received from the
CM11A just before a triple 0xff sequence indicating an outgoing command
or an RF signal or whatever else is appended to the spool file. In this
rare case, the fourth 0xff detected is treated as incorrect sequence
length, resulting in the whole sequence being miss-interpreted and the
just missed checksum still expected to be catched.

Detect this condition and process correctly, invalidating the expected
checksum value and processing the remaining triple 0xff initiated
sequence.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>