Link with Autoconf selected libraries
Instead of specifying a list of libraries to link the Heyu binary with,
depend on Autoconf library detection features.

Since libucb, libgen and libcmd libraries, supposed to be required on
the ATT SVR4 platform, are not supported by configure.ac, I've left
them specified explicitly, but passed to ./configure to be substituted
from Makefile.in instead of hardcoded into config.mk.

No changes to Makefile.am required.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>