Don't freopen() closed streams
While starting the engine, previously close()d stdout and stderr
streams can be freopen()ed to point to the log file. This may lead to
unpredictable behaviour, like the engine no longer reporting events to
the log, observed on recent Fedora 15 installations. See #18.

Fix this by dropping initial close()s of fd's 1 and 2, which will now
end up being redirected to /dev/null if a log file is not used, instead
of closed.

Created against Heyu 2.9.4
Tested on Fedora 15, as well as on not yet affected Gentoo/Linux and
OpenSolaris 0906.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>