Avoid using custom ATTSVR4 build time macro
There is a macro, ATTSVR4, set with the Heyu Configure.sh script when
an ATT flavour of a UNIX system is detected. This macro is then used at
compile time for conditional inclusion of some header files, as well as
for selecting usleep() as a replacement for nanosleep(), which is
supposed to be missing.

Since all conditionally included header files have already their
Autoconf based conditions in place, there is no need to use the
ATTSVR4 for this purpose any longer. However, there are cases where
this macro is used, together with others, in #ifdef..#else..#endif
constructs, which may require a different approach, so we leave them
untouched for now.

For selecting usleep() over nanosleep(), we replace this macro with
a negation of the one generated with Autoconf check for nanosleep()
availability.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>