Let Autoconf create Makefile from Makefile.in
Further migration to Autotools requires activation of Makefile auto-
generation feature of Autoconf, with Makefile.in serving as a template.

Instead of creating the top of Makefile by Configure.sh script and
appending its bottom directly from Makefile.in, put the top into a
separate config.mk file and include it from the Makefile.in. Then,
allow configure to create a Makefile from the Makefile.in template.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>