Switch over to using Automake
After all recent changes, it should be now possible to configure, build
and install Heyu using an Automake generated Makefile.in. The Automake
initialization directive has been uncommented in configure.ac, and
calls to autoconf and autoheader replaced with autoreconf in
Configure.sh.

To give it a try, just install an Automake package in addition to
Autoconf and follow the Heyu installation procedure as usually.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>