Revert "Compare RFXCOM X10 signal bursts by type, not by length"
This reverts commit a65f3b9a1a0fbad599f4b528a0f6ddce2f4f10be.

While working on Oregon radio clocks support, it occurred that still
more advances methods of suppressing repeated bursts of same signals
have to be invented.  Instead of implementing them on top of the
recently relaxed check for signals similarity, revert back to the
previous, more strict method, with the intention of further extending
it with case specific exceptions.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>