Engine: don't handle checksums when processing RF signals
Commit 2cd54bdeb9, "Engine: use chksum_alert for all CM11 commands",
introduced a mechanism for calculating checksums of all commands
destined to a power line interface and passing them to the procedure
that reads data back from that interface, in order to detect and ignore
all checksum responses instead of reporting them as unknown values.
However, that change didn't take into account that the code path which
handled outgoing power line commands was also used for processing of
X10 Standard RF signals obtained from an AUX interface even if not
retransmitted over the power line. As a consequence, it may happen that
an excessively stored checksum of an incoming RF signal will match an
arbitrary byte of data read from the power line interface, resulting in
temporary misinterpretation of the incoming data stream. Fix it.

Created against heyu-2.10-rc2.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>