Prepare Heyu build system for xPL integration
Before we can extend Heyu source code with xPL support, we should be
able to build Heyu in an environment where most probably no xPL headers
nor libraries are pre-installed.

This patch extends our Autotools configuration not only with checks for
xPL support files availability, but also with a setup that results in
an attempt to build libxPL.a automatically if not already installed.

No functional changes to Heyu itself.

Somewhat massive changes to configure, config.h.in and Makefile.in have
been generated automatically with autoreconf from updated configure.ac
and Makefile.am sources.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>