X10 Security: use 16-bit IDs by default if xPL enabled
Support for 16-bit X10 Security module IDs could be activated only if
an RFXCOM receiver in variable length mode was configured on a TTY_AUX
port. This limitation should not apply if Heyu is build with xPL
support enabled, or we are not able to handle 16-bit IDs from xPL
x10.security messages.

Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>