Don't test for malloc/realloc(0) GNU libc compatibility
When cross-compiling, an assumption is made by autoconf that behaviour
of malloc/realloc(0) is not compatible with GNU libc, and the build
breaks because of replacement functions rpl_malloc/rpl_realloc()
missing.

Instead for requiring the autoscan suggested GNU libc compatibility of
malloc/realloc(0) behaviour, prevent from ever passing 0 size to
malloc/realloc() and replace the autoscan generated tests with simpler
tests for just availability of the functions.

Changes to config.h.in and configure are autoreconf generated.

Reported-by: <dony71@gmail.com>
Signed-off-by: Janusz Krzysztofik <jkrzyszt@tis.icnet.pl>